+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [ENHANCEMENT] Added link to active alerts directly from providers
 * [ENHANCEMENT] Added POJO support in MongoDB client
 * [ENHANCEMENT] Added configuration option for `ALERTMONITOR_MONGODB_CONNECT_TIMEOUT_SEC` and `ALERTMONITOR_MONGODB_READ_TIMEOUT_SEC`
+* [ENHANCEMENT] Added `provider` attribute to alerts
 
 ## 2.4.4-SNAPSHOT